+++ Users.js
@@ -17,7 +17,7 @@ import FilterControlGroup from '@folio/stripes-components/lib/FilterControlGroup
 import Layer from '@folio/stripes-components/lib/Layer';
 import FilterGroups, { initialFilterState, onChangeFilter } from '@folio/stripes-components/lib/FilterGroups';
 import transitionToParams from '@folio/stripes-components/util/transitionToParams';
-import makePathFunction from '@folio/stripes-components/util/makePathFunction';
+import makeQueryFunction from '@folio/stripes-components/util/makeQueryFunction';
 import IfPermission from '@folio/stripes-components/lib/IfPermission';
 
 import UserForm from './UserForm';
@@ -56,6 +56,7 @@ class Users extends React.Component {
         log: PropTypes.func.isRequired,
       }).isRequired,
       connect: PropTypes.func.isRequired,
+      hasPerm: PropTypes.func.isRequired,
     }).isRequired,
     data: PropTypes.object.isRequired,
     history: PropTypes.shape({
@@ -85,20 +86,24 @@ class Users extends React.Component {
       records: 'users',
       recordsRequired: 30,
       perRequest: 10,
-      path: makePathFunction(
-        'users',
-        'username=*',
-        'username="$QUERY*" or personal.first_name="$QUERY*" or personal.last_name="$QUERY*"',
-        {
-          Active: 'active',
-          Name: 'personal.last_name personal.first_name',
-          'Patron Group': 'patron_group',
-          'User ID': 'username',
-          Email: 'personal.email',
+      path: 'users',
+      GET: {
+        params: {
+          query: makeQueryFunction(
+            'username=*',
+            'username="$QUERY*" or personal.first_name="$QUERY*" or personal.last_name="$QUERY*"',
+            {
+              Active: 'active',
+              Name: 'personal.last_name personal.first_name',
+              'Patron Group': 'patron_group',
+              'User ID': 'username',
+              Email: 'personal.email',
+            },
+            filterConfig,
+          ),
         },
-        filterConfig,
-      ),
-      staticFallback: { path: 'users' },
+        staticFallback: { params: {} },
+      },
     },
     patronGroups: {
       type: 'okapi',
@@ -202,6 +207,7 @@ class Users extends React.Component {
   }
 
   postCreds(username, creds) {
+    this.log('xhr', `POST credentials for new user '${username}':`, creds);
     fetch(`${this.okapi.url}/authn/credentials`, {
       method: 'POST',
       headers: Object.assign({}, { 'X-Okapi-Tenant': this.okapi.tenant, 'X-Okapi-Token': this.okapi.token, 'Content-Type': 'application/json' }),
@@ -210,12 +216,18 @@ class Users extends React.Component {
       if (response.status >= 400) {
         this.log('xhr', 'Users. POST of creds failed.');
       } else {
-        this.postPerms(username, ['users.read', 'usergroups.read', 'perms.permissions.get']);
+        this.postPerms(username, [
+          'users.collection.get',       // so the user can search for his own user record after login
+          'perms.permissions.get',      // so the user can fetch his own permissions after login
+          'usergroups.collection.get',  // so patron groups can be listed in the Users module
+          'module.trivial.enabled',     // so that at least one module is available to new users
+        ]);
       }
     });
   }
 
   postPerms(username, perms) {
+    this.log('xhr', `POST permissions for new user '${username}':`, perms);
     fetch(`${this.okapi.url}/perms/users`, {
       method: 'POST',
       headers: Object.assign({}, { 'X-Okapi-Tenant': this.okapi.tenant, 'X-Okapi-Token': this.okapi.token, 'Content-Type': 'application/json' }),
@@ -262,14 +274,39 @@ class Users extends React.Component {
       Email: user => _.get(user, ['personal', 'email']),
     };
 
+    const detailsPane = (
+      this.props.stripes.hasPerm('users.item.get') ?
+        (<Route
+          path={`${this.props.match.path}/view/:userid/:username`}
+          render={props => <this.connectedViewUser stripes={stripes} paneWidth="44%" onClose={this.collapseDetails} {...props} />}
+        />) :
+        (<div
+          style={{
+            position: 'absolute',
+            right: '1rem',
+            bottom: '1rem',
+            width: '34%',
+            zIndex: '9999',
+            padding: '1rem',
+            backgroundColor: '#fff',
+          }}
+        >
+          <h2>Permission Error</h2>
+          <p>Sorry - your user permissions do not allow access to this page.</p>
+        </div>));
+
     return (
       <Paneset>
         {/* Filter Pane */}
         <Pane defaultWidth="16%" header={searchHeader}>
           <FilterGroups config={filterConfig} filters={this.state.filters} onChangeFilter={this.onChangeFilter} />
           <FilterControlGroup label="Actions">
-            <IfPermission {...this.props} perm="users-bl.createuser">
-              <Button fullWidth onClick={this.onClickAddNewUser}>New user</Button>
+            <IfPermission {...this.props} perm="users.item.post">
+              <IfPermission {...this.props} perm="login.item.post">
+                <IfPermission {...this.props} perm="perms.users.item.post">
+                  <Button fullWidth onClick={this.onClickAddNewUser}>New user</Button>
+                </IfPermission>
+              </IfPermission>
             </IfPermission>
           </FilterControlGroup>
         </Pane>
@@ -300,11 +337,7 @@ class Users extends React.Component {
           />
         </Pane>
 
-        {/* Details Pane */}
-        <Route
-          path={`${this.props.match.path}/view/:userid/:username`}
-          render={props => <this.connectedViewUser stripes={stripes} paneWidth="44%" onClose={this.collapseDetails} {...props} />}
-        />
+        {detailsPane}
         <Layer isOpen={data.addUserMode ? data.addUserMode.mode : false} label="Add New User Dialog">
           <UserForm
             initialValues={{ available_patron_groups: this.props.data.patronGroups }}